'student'

+++ resources/views/project/admin/users/ajax/supervisorMajorList.blade.php
@@ -14,9 +14,11 @@
                                 <h5 class="mb-0 counter">{{$key->majors->m_name}}</h5>
                                 <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-shopping-bag icon-bg"><path d="M6 2L3 6v14a2 2 0 0 0 2 2h14a2 2 0 0 0 2-2V6l-3-4z"></path><line x1="3" y1="6" x2="21" y2="6"></line><path d="M16 10a4 4 0 0 1-8 0"></path></svg>
                             </div>
-                            <div class="col-md-2">
-                                <button class="btn btn-lg" onclick="delete_major_for_supervisor({{$key->ms_id}})" type="button"><span class="fa fa-trash "></span></button>
-                            </div>
+                            @if (auth()->user()->u_role_id == 1) {{-- Admin --}}
+                                <div class="col-md-2">
+                                    <button class="btn btn-lg" onclick="delete_major_for_supervisor({{$key->ms_id}})" type="button"><span class="fa fa-trash "></span></button>
+                                </div>
+                            @endif
                         </div>
                     </div>
                 </div>
